requirements.txt
gradio
openai-whisper
translate
langdetect
google-cloud-texttospeech